+++ doc/source/changelog.rst
@@ -4,7 +4,8 @@ Changelog
 next
 ----
 * Bumped docutils dependency to >=0.12, for setuptools compatibility.
-* Added support for Python 3.10.
+* Added support for Python 3.10, and for ``A | B``-style unions (only on that
+  version of Python).
 * Fixed bad interaction between custom None parsers and tuple parsers.
 * Give preference to None parsers in Unions.
 * `defopt.signature` annotates the return type with the documented raisable